JavaDoc, Utitlities, New Basic Features
We added JavaDoc into every class on every public method. Furthemore, we
created a Math Util for convert millis to ticks and ticks to millis. We
also improved the Bukkit Scheduler API with a java.util.Timer-like
ScheduleGroup and ScheduleTask and created an Handling API for the
Bukkit-Thread-Problem (you can call Bukkit-API-functions only in the
Main-Bukkit-Thread).