Exception-System, Command-System, Fixes
We added much stuff in this big update. At first, we fixed some small
bugs in the code.
We also integrated an Exception-System for exceptions like somebody
hasn't enough permissions to execute a command. With the exception you
can maybe output high-personalized error messages.
An other big thing is the new Command-System. You can simply define
commands like /command help and customize them on a high level. It's
very easy to use, but it's fast and modern.