Update to latest Bukkit-Build
We updated QuarterBukkit to the latest Bukkit-Build 1.4.6-R0.1. By doing
this, we recognized that we can't provide the total compatibility to
other versions with our Name-Tag-API. We don't want to add some
version-import-stuff and lose the compatibility, so we decided to remove
this feature temporarily. We're sorry. If you really want to use this
feature, you can use the Tag-API
(http://dev.bukkit.org/server-mods/tag/)
Furthemore, we now provide compatibility to the latest build on every
other class. Hoping you don't miss our Name-Tag-Feature :)