Add ActiveObjectSystem-awareness to BaseObject
Previously, most of the users and -- in some cases -- even the object
classes themselves had no clue about the context of an object and were
instead faced with just the object on it's own. Now, however, each
object stores the ActiveObjectSystem it is part of and makes that
variable accessible to everyone.